Finished all
all entities

+++ src/main/java/com/babyboom/dtos/Enfermedad_bebeDTO.java
@@ -9,15 +9,16 @@ import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 
 public class Enfermedad_bebeDTO {
-    private Long id;
+    private int id;
     private TipoEnfermedad tipoEnfermedad;
     private Bebe bebe;
     private String sintonmas;
-    public Long getId() {
+
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
