+++ src/main/java/com/babyboom/servicesimplement/Enfermedad_bebeServiceImplement.java
@@ -27,13 +27,13 @@ public class Enfermedad_bebeServiceImplement implements IEnfermedad_BebeService 
     }
 
     @Override
-    public void delete(long id) {
-        fR.deleteById((int) id);
+    public void delete(int id) {
+        fR.deleteById(id);
 
     }
 
     @Override
-    public Enfermedad_bebe listId(long id) {
-        return fR.findById((int) id).orElse(new Enfermedad_bebe());
+    public Enfermedad_bebe listId(int id) {
+        return fR.findById(id).orElse(new Enfermedad_bebe());
     }
 }
